move util to utils file
