+++ packages/node-opentelemetry/src/otel.ts
@@ -49,6 +49,7 @@ import { getHyperDXMetricReader } from './metrics';
 import { MutableAsyncLocalStorageContextManager } from './MutableAsyncLocalStorageContextManager';
 import { Logger as OtelLogger } from './otel-logger';
 import HyperDXSpanProcessor from './spanProcessor';
+import { parseOtlpHeaders } from './utils';
 
 const UI_LOG_PREFIX = '[⚡HyperDX]';
 
@@ -144,43 +145,6 @@ const isSupported = (
 
 const hrtimeToMs = (hrtime: [number, number]) => {
   return hrtime[0] * 1e3 + hrtime[1] / 1e6;
-};
-
-/**
- * Parses OTEL_EXPORTER_OTLP_HEADERS environment variable into a structured headers object.
- * Format: "key1=value1,key2=value2" -> { key1: "value1", key2: "value2" }
- */
-export const parseOtlpHeaders = (
-  headersString?: string,
-): Record<string, string> => {
-  if (!headersString) {
-    return {};
-  }
-
-  const headers: Record<string, string> = {};
-  const pairs = headersString.split(',');
-
-  for (const pair of pairs) {
-    const trimmedPair = pair.trim();
-    if (!trimmedPair) {
-      continue;
-    }
-
-    const equalIndex = trimmedPair.indexOf('=');
-    if (equalIndex === -1) {
-      // Skip malformed pairs without '='
-      continue;
-    }
-
-    const key = trimmedPair.substring(0, equalIndex).trim();
-    const value = trimmedPair.substring(equalIndex + 1).trim();
-
-    if (key) {
-      headers[key] = value;
-    }
-  }
-
-  return headers;
 };
 
 const healthCheckUrl = async (
